Refactor splunkhttp config processing
