+++ instrumentation/net/http/splunkhttp/config.go
@@ -26,46 +26,31 @@ const (
 	envVarTraceResponseHeaderEnabled = "SPLUNK_TRACE_RESPONSE_HEADER_ENABLED" // Adds `Server-Timing` header to HTTP responses
 )
 
-// WithOTelOpts is use to pass the OpenTelemetry SDK options.
-func WithOTelOpts(opts ...otelhttp.Option) Option {
-	return optionFunc(func(cfg *config) {
-		cfg.OTelOpts = opts
-	})
-}
-
 // WithTraceResponseHeader enables or disables the TraceResponseHeaderMiddleware.
 //
 // The default is to enable the TraceResponseHeaderMiddleware if this option is not passed.
 // Additionally, the SPLUNK_TRACE_RESPONSE_HEADER_ENABLED environment variable
 // can be set to TRUE or FALSE to specify this option. This option value will be
 // given precedence if both it and the environment variable are set.
-func WithTraceResponseHeader(v bool) Option {
-	return optionFunc(func(cfg *config) {
+func WithTraceResponseHeader(v bool) otelhttp.Option {
+	return optionWrapper{apply: func(cfg *config) {
 		cfg.TraceResponseHeaderEnabled = v
-	})
+	}}
 }
 
-// Option is used for setting optional config properties.
-type Option interface {
-	apply(*config)
-}
-
-// config represents the available configuration options.
+// config represents the available Splunk specific configuration options.
 type config struct {
-	OTelOpts                   []otelhttp.Option
 	TraceResponseHeaderEnabled bool
 }
 
-// optionFunc provides a convenience wrapper for simple Options
-// that can be represented as functions.
-type optionFunc func(*config)
-
-func (o optionFunc) apply(c *config) {
-	o(c)
+// optionWrapper provides a convenience wrapper for Splunk specific options.
+type optionWrapper struct {
+	otelhttp.Option
+	apply func(*config)
 }
 
-// newConfig creates a new config struct and applies opts to it.
-func newConfig(opts ...Option) *config {
+// newConfig creates a new config struct and applies Splunk specific options to it.
+func newConfig(opts ...otelhttp.Option) *config {
 	traceResponseHeaderEnabled := true
 	if v := os.Getenv(envVarTraceResponseHeaderEnabled); strings.EqualFold(v, "false") {
 		traceResponseHeaderEnabled = false
@@ -75,7 +60,10 @@ func newConfig(opts ...Option) *config {
 		TraceResponseHeaderEnabled: traceResponseHeaderEnabled,
 	}
 	for _, opt := range opts {
-		opt.apply(c)
+		// apply only Splunk specific options
+		if splunkOpt, ok := opt.(optionWrapper); ok {
+			splunkOpt.apply(c)
+		}
 	}
 	return c
 }
