+++ instrumentation/net/http/splunkhttp/example_test.go
@@ -30,7 +30,7 @@ func ExampleTraceResponseHeaderMiddleware() {
 	var handler http.Handler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		io.WriteString(w, "Hello world") //nolint:errcheck
 	})
-	handler = splunkhttp.NewHandler(handler, "server", splunkhttp.WithOTelOpts(otelhttp.WithTracerProvider(oteltest.NewTracerProvider())))
+	handler = splunkhttp.NewHandler(handler, "server", otelhttp.WithTracerProvider(oteltest.NewTracerProvider()))
 
 	ts := httptest.NewServer(handler)
 	defer ts.Close()
