+++ instrumentation/net/http/splunkhttp/trace_response_header_test.go
@@ -23,12 +23,13 @@ import (
 	"go.opentelemetry.io/otel/oteltest"
 )
 
-func TestNewHandlerTraceResponseHeaderDisabled(t *testing.T) {
+func TestTraceResponseHeaderMiddleware(t *testing.T) {
 	resp := responseForHandler(func(handler http.Handler) http.Handler { // nolint:bodyclose // Body is not used
-		return NewHandler(handler, "server", WithOTelOpts(otelhttp.WithTracerProvider(oteltest.NewTracerProvider())), WithTraceResponseHeader(false))
+		handler = TraceResponseHeaderMiddleware(handler)
+		return otelhttp.NewHandler(handler, "server", otelhttp.WithTracerProvider(oteltest.NewTracerProvider()))
 	})
 
 	assert.Equal(t, http.StatusOK, resp.StatusCode, "should return OK status code")
-	assert.NotContains(t, resp.Header["Access-Control-Expose-Headers"], "Server-Timing", "should NOT set Access-Control-Expose-Headers header")
-	assert.NotRegexp(t, "^traceparent;desc=\"00-[0-9a-f]{32}-[0-9a-f]{16}-01\"$", resp.Header.Get("Server-Timing"), "should not add traceID to Server-Timing header")
+	assert.Contains(t, resp.Header["Access-Control-Expose-Headers"], "Server-Timing", "should set Access-Control-Expose-Headers header")
+	assert.Regexp(t, "^traceparent;desc=\"00-[0-9a-f]{32}-[0-9a-f]{16}-01\"$", resp.Header.Get("Server-Timing"), "should return properly formated Server-Timing header")
 }
